+++ La-Mula-App-main/src/main/kotlin/com/lamulaapp/controller/dto/ProductDto.kt
@@ -1,5 +1,6 @@
 package com.lamulaapp.controller.dto
 
+import java.math.BigDecimal
 import java.util.UUID
 
 data class ProductDto(
@@ -7,6 +8,6 @@ data class ProductDto(
     val name: String,
     val idCoffeeType: UUID,  // Usaremos el ID del CoffeeType en el DTO
     val description: String? = null,
-    val unitCost: Double,
+    val unitCost: BigDecimal,
     val quantity: Int
 )
